Update version2.py
This uses the chosen component of outputting whether each answer is correct or not, before moving on to the next question.

The quiz is functional - random questions are generated, the questions displayed, the answers processed, the score calculated, with the option to view the response and export it.

Has some minor errors however - i.e. the score is always zero. No error handling yet for the answer entries.